+++ sdk/monitor/azquery/models.go
@@ -82,7 +82,7 @@ type Body struct {
 
 	// Optional. The timespan over which to query data. This is an ISO8601 time period value. This timespan is applied in addition
 	// to any that are specified in the query expression.
-	Timespan *ISO8601TimeInterval `json:"timespan,omitempty"`
+	Timespan *TimeInterval `json:"timespan,omitempty"`
 }
 
 // Column - A column in a table.
@@ -110,12 +110,8 @@ type LogsClientQueryBatchOptions struct {
 
 // LogsClientQueryWorkspaceOptions contains the optional parameters for the LogsClient.QueryWorkspace method.
 type LogsClientQueryWorkspaceOptions struct {
-	// Optional. The prefer header to set server timeout, query statistics and visualization information. For server timeout,
-	// set the prefer string equal to "wait=NumOfMinutes", where NumOfMinutes is an
-	// integer (ex 600). Max server time is ten minutes. For statistics, set prefer string equal to "include-statistics=true".
-	// For visualization, set prefer string equal to "include-render=true". For more
-	// information, see https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/monitor/azquery#readme-increase-wait-time-include-statistics-include-render-visualization
-	Prefer *string
+	// Optional. The prefer header to set server timeout, query statistics and visualization information.
+	Options *LogsQueryOptions
 }
 
 // MetadataValue - Represents a metric metadata value.
@@ -304,7 +300,7 @@ type MetricsClientQueryResourceOptions struct {
 	// Reduces the set of data collected. The syntax allowed depends on the operation. See the operation's description for details.
 	ResultType *ResultType
 	// The timespan of the query. It is a string with the following format 'startDateTimeISO/endDateTimeISO'.
-	Timespan *ISO8601TimeInterval
+	Timespan *TimeInterval
 	// The maximum number of records to retrieve. Valid only if $filter is specified. Defaults to 10.
 	Top *int32
 }
@@ -314,7 +310,7 @@ type Response struct {
 	// REQUIRED; The timespan for which the data was retrieved. Its value consists of two datetimes concatenated, separated by
 	// '/'. This may be adjusted in the future and returned back from what was originally
 	// requested.
-	Timespan *ISO8601TimeInterval `json:"timespan,omitempty"`
+	Timespan *TimeInterval `json:"timespan,omitempty"`
 
 	// REQUIRED; the value of the collection.
 	Value []*Metric `json:"value,omitempty"`